prefetch: fix replay_test test error
The `copy_uncached_files_and_record_from` test utility function is
updated to keep the file descriptors of the newly created "uncached"
files open and return them.

This change ensures that the files are not closed for the duration of
the test by returning a vector of open `File` handles. In some certain
environment, the closed file could be readed by other processes for
security check, which will cause unexpected test failure.

Bug: 436685355
Test: cargo build & cargo test

Change-Id: Ib8eb590c7ab0d0c84369f2845e9d1ecca4b3de45

Former-commit-id: 7a8f860b30ac9859e8427ebe073940d3893f7954