Fix recovery sideload failing if stale snapshots are present.
Recovery attempts to call MapAllSnapshots() during filesystem verification. If there happens to be a leftover snapshot from a previous OTA, MapAllSnapshots() can fail. Restrict it to only map snapshots that are part of the current OTA.

Note that normally DeleteSnapshot would clean these up, but we explicitly don't do that in recovery (maybe we should).

Bug: 438092240
Test: RecoveryOtaTest
Change-Id: I4a7cf7a38c20ff7e00a215c25e49004a1d76eb49

Former-commit-id: 1466c7cd1daf5405c8f70d5cb2a118d259a22984